version2.4.1 修复用户退出时不检查用户状态的bug 增加限制：项目统计图只有在选择多个学院的时候才显示 修复界面bug：人员信息所选学院中错误限制col为3，已修改为12 事务同步管理：为People,Project,Paper，Patent,Award,Lab的import增加事务过程 修复bug：paper列表里通讯作者直接显示姓名，不通过id 修复bug：一作的id为88888时不显示链接 增加限制：列表显示的项目序号不直接使用id而是使用ng-repeat的$index+1来显示当前状态下的项目真实序列号     vm.baseIndex = (vm.paginationConf.currentPage-1) * vm.paginationConf.itemsPerPage;     id = vm.baseIndex+$index+1 调整people表中各列比例，增加一级学科和二级学科宽度

+++ EXPERT/Home/Controller/AwardController.class.php
@@ -374,28 +374,37 @@ class AwardController extends Controller {
                         $condition["achievement"] =$data["achievement"];
                         $condition["name"]=$data["name"];
                         $isduplicated = $award->where($condition)->find();
+                        $award->startTrans();
                         if((int)$isduplicated['id']>0){//数据库中存在相同数据，使用更新操作
                             $num = $award->where($condition)->save($data);
-                            $result = $isduplicated['id'];
-                            //将成功更新的数据记录到日志中
-                            $update_counter++;
-                            $updated_id[]=$result;
+                            if($num){
+                                $award->commit();
+                                $result = $isduplicated['id'];
+                                //将成功更新的数据记录到日志中
+                                $update_counter++;
+                                $updated_id[]=$result;
 
-                            //删除所有award_other_person相关award职工号，稍后重新插入（以后来数据为准）
-                            $flag["award_id"]=(int)$result;
-                            M('award_other_person')->where($flag)->delete();
+                                //删除所有award_other_person相关award职工号，稍后重新插入（以后来数据为准）
+                                $flag["award_id"]=(int)$result;
+                                M('award_other_person')->where($flag)->delete();
+                            }
 
                         }else{
                             //插入数据
                             $result = $award->add($data);
-                            //成功插入数据记录到日志中
-                            $insert_counter++;
-                            $inserted_id[]=$result;
+                            if($result){
+                                $award->commit();
+                                //成功插入数据记录到日志中
+                                $insert_counter++;
+                                $inserted_id[]=$result;
+                            }
+
                         }
 
 
                         if (empty($result))
                         {
+                            $award->rollback();
                             $error_counter++;
                             $errored_name[]=$data["name"];
                         }else if($others_names!=''){
