+++ EXPERT/Home/Controller/ProjectController.class.php
@@ -433,26 +433,35 @@ class ProjectController extends Controller {
                         $condition["unique_code"]=$data["unique_code"];
                         $isduplicated = $project->where($condition)->find();
 //                        $isduplicated = $project->where($condition)->select();//使用name做检索时此处使用find会出现bug，需要使用select
+                        $project->startTrans();
 
                         if((int)$isduplicated['id']>0){//数据库中存在相同数据，使用更新操作
                             $num = $project->where($condition)->save($data);
-                            $result = $isduplicated['id'];
-                            //将成功更新的数据记录到日志中
-                            $update_counter++;
-                            $updated_id[]=$result;
+                            if($num){
+                                $project->commit();
+                                $result = $isduplicated['id'];
+                                //将成功更新的数据记录到日志中
+                                $update_counter++;
+                                $updated_id[]=$result;
 
-                            //删除所有project_other_persons相关表，稍后重新插入（以后来数据为准）
-                            $flag["project_id"]=(int)$result;
-                            M('project_other_persons')->where($flag)->delete();
+                                //删除所有project_other_persons相关表，稍后重新插入（以后来数据为准）
+                                $flag["project_id"]=(int)$result;
+                                M('project_other_persons')->where($flag)->delete();
+                            }
+
                         }else{
                             //插入数据
                             $result = $project->add($data);
                             //成功插入数据记录到日志中
-                            $insert_counter++;
-                            $inserted_id[]=$result;
+                            if($result){
+                                $project->commit();
+                                $insert_counter++;
+                                $inserted_id[]=$result;
+                            }
                         }
                         if (empty($result))
                         {
+                            $project->rollback();
                             $error_counter++;
                             $errored_name[]=$data["name"]."(插入失败)";
                         }else if($expert_participants!=null){
